Splitted create observable component
